adjust test TextFrameTests
