--- test/ShapeCrawler.Tests.Unit.xUnit/Helpers/StreamExtensions.cs
@@ -1,13 +1,0 @@
-﻿using System.IO;
-
-namespace ShapeCrawler.Tests.Unit.Helpers;
-
-public static class StreamExtensions
-{
-    public static void ToFile(this Stream stream, string filePath)
-    {
-        stream.Position = 0;
-        using var destStream = new FileStream(filePath, FileMode.Create, FileAccess.Write);
-        stream.CopyTo(destStream);
-    }
-}
