+++ test/ShapeCrawler.Tests.Unit.xUnit/Helpers/TestCase.cs
@@ -64,46 +64,6 @@ public class TestCase
     }
 }
 
-public class TestCase<T1>
-{
-    private readonly int testCaseNumber;
-    private IPresentation pres;
-    public T1 Param1 { get; }
-    public IPresentation Presentation
-    {
-        get
-        {
-            if (this.pres != null)
-            {
-                return this.pres;
-            }
-            
-            var stream = TestHelper.GetStream(this.PresentationName);
-            this.pres = SCPresentation.Open(stream);
-
-            return this.pres;
-        }
-    }
-
-    public string PresentationName { get; set; }
-
-    public TestCase(int testCaseNumber, T1 param1)
-    {
-        this.testCaseNumber = testCaseNumber;
-        this.Param1 = param1;
-    }
-
-    public override string ToString()
-    {
-        return this.testCaseNumber.ToString(NumberFormatInfo.CurrentInfo);
-    }
-
-    public void SetPresentation(MemoryStream mStream)
-    {
-        throw new System.NotImplementedException();
-    }
-}
-
 public class TestCase<T1, T2>
 {
     private readonly int testCaseNumber;
