+++ test/ShapeCrawler.Tests.Unit.xUnit/TextFrameTests.cs
@@ -1,16 +1,10 @@
-#if DEBUG
-
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Linq;
 using FluentAssertions;
-using NUnit.Framework;
-using ShapeCrawler.Shapes;
-using ShapeCrawler.Tests.Shared;
-using ShapeCrawler.Tests.Unit.xUnit.Helpers;
-using ShapeCrawler.Tests.Unit.xUnit.Helpers.Attributes;
+using ShapeCrawler.Tests.Unit.Helpers;
+using ShapeCrawler.Tests.Unit.Helpers.Attributes;
 using Xunit;
 
 // ReSharper disable All
@@ -22,42 +16,15 @@ namespace ShapeCrawler.Tests.Unit.xUnit
     [SuppressMessage("Usage", "xUnit1013:Public method should be marked as test")]
     public class TextFrameTests : SCTest
     {
-        [Fact]
-        public void Text_Getter_returns_text_of_table_Cell()
-        {
-            // Arrange
-            var pptx8 = GetInputStream("008.pptx");
-            var pres8 = SCPresentation.Open(pptx8);
-            var pptx1 = GetInputStream("001.pptx");
-            var pres1 = SCPresentation.Open(pptx1);
-            var pptx9 = GetInputStream("009_table.pptx");
-            var pres9 = SCPresentation.Open(pptx9);
-            var textFrame1 = ((IAutoShape)SCPresentation.Open(GetInputStream("008.pptx")).Slides[0].Shapes.First(sp => sp.Id == 3)).TextFrame;
-            var textFrame2 = ((ITable)SCPresentation.Open(GetInputStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
-                .TextFrame;
-            var textFrame3 = ((ITable)SCPresentation.Open(GetInputStream("009_table.pptx")).Slides[2].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
-                .TextFrame;
-            
-            // Act
-            var text1 = textFrame1.Text;
-            var text2 = textFrame2.Text;
-            var text3 = textFrame3.Text;
-
-            // Act
-            text1.Should().NotBeEmpty();
-            text2.Should().BeEquivalentTo("id3");
-            text3.Should().BeEquivalentTo($"0:0_p1_lvl1{Environment.NewLine}0:0_p2_lvl2");
-        }
-        
         [Xunit.Theory]
-        // [SlideShapeData("009_table.pptx", 4, 2, "Title text")]
-        // [SlideShapeData("001.pptx", 1, 5, " id5-Text1")]
-        // [SlideShapeData("019.pptx", 1, 2, "1")]
-        // [SlideShapeData("014.pptx", 2, 5, "Test subtitle")]
-        // [SlideShapeData("011_dt.pptx", 1, 54275, "Jan 2018")]
-        // [SlideShapeData("021.pptx", 4, 2, "test footer")]
-        // [SlideShapeData("012_title-placeholder.pptx", 1, 2, "Test title text")]
-        // [SlideShapeData("012_title-placeholder.pptx", 1, 3, "P1 P2")]
+        [SlideShapeData("009_table.pptx", 4, 2, "Title text")]
+        [SlideShapeData("001.pptx", 1, 5, " id5-Text1")]
+        [SlideShapeData("019.pptx", 1, 2, "1")]
+        [SlideShapeData("014.pptx", 2, 5, "Test subtitle")]
+        [SlideShapeData("011_dt.pptx", 1, 54275, "Jan 2018")]
+        [SlideShapeData("021.pptx", 4, 2, "test footer")]
+        [SlideShapeData("012_title-placeholder.pptx", 1, 2, "Test title text")]
+        [SlideShapeData("012_title-placeholder.pptx", 1, 3, "P1 P2")]
         [SlideShapeData("031.pptx", 46, 5, "P3")]
         public void Text_Getter_returns_text(IShape shape, string expectedText)
         {
@@ -69,25 +36,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
 
             // Assert
             text.Should().BeEquivalentTo(expectedText);
-        }
-        
-        // [SlideShapeData("031.pptx", 46, 5, "P3")]
-        [Fact]
-        public void Text_Getter_returns_text_from_New_Slide()
-        {
-            throw new Exception("In Progress.");
-            // Arrange
-            var pptx = GetInputStream("031.pptx");
-            var pres = SCPresentation.Open(pptx);
-            var layout = pres.SlideMasters[0].SlideLayouts[0];
-
-            // Act
-            pres.Slides.AddEmptySlide(layout);
-            var textFrame = pres.Slides.Last().Shapes.GetById<IAutoShape>(5).TextFrame;
-            var text = textFrame.Text;
-            
-            // Assert
-            text.Should().BeEquivalentTo("some");
         }
 
         [Xunit.Theory]
@@ -133,27 +81,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
             }
         }
 
-        [Fact]
-        public void Text_Setter_can_update_content_multiple_times()
-        {
-            // Arrange
-            var pptx = GetInputStream("autoshape-case005_text-frame.pptx");
-            var pres = SCPresentation.Open(pptx);
-            var textFrame = pres.Slides[0].Shapes.GetByName<IAutoShape>("TextBox 1").TextFrame;
-            var modifiedPres = new MemoryStream();
-
-            // Act
-            textFrame.Text = textFrame.Text.Replace("{{replace_this}}", "confirm this");
-            textFrame.Text = textFrame.Text.Replace("{{replace_that}}", "confirm that");
-
-            // Assert
-            pres.SaveAs(modifiedPres);
-            pres.Close();
-            pres = SCPresentation.Open(modifiedPres);
-            textFrame = pres.Slides[0].Shapes.GetByName<IAutoShape>("TextBox 1").TextFrame;
-            textFrame.Text.Should().ContainAll("confirm this", "confirm that");
-        }
-        
         [Xunit.Theory]
         [MemberData(nameof(TestCasesTextSetter))]
         public void Text_Setter_updates_content(TestElementQuery testTextBoxQuery)
@@ -230,44 +157,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
             }
         }
 
-        [Fact]
-        public void Text_Setter_updates_text_box_content_and_Reduces_font_size_When_text_is_Overflow()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("001.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var textBox = pres.Slides[0].Shapes.GetByName<IAutoShape>("TextBox 8");
-            var textFrame = textBox.TextFrame;
-            var fontSizeBefore = textFrame.Paragraphs[0].Portions[0].Font.Size;
-            var newText = "Shrink text on overflow";
-
-            // Act
-            textFrame.Text = newText;
-
-            // Assert
-            textFrame.Text.Should().BeEquivalentTo(newText);
-            textFrame.Paragraphs[0].Portions[0].Font.Size.Should().Be(8);
-        }
-        
-        [Fact]
-        public void Text_Setter_resizes_shape_to_fit_text()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("autoshape-case003.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var shape = pres.Slides[0].Shapes.GetByName<IAutoShape>("AutoShape 4");
-            var textFrame = shape.TextFrame;
-
-            // Act
-            textFrame.Text = "AutoShape 4 some text";
-
-            // Assert
-            shape.Height.Should().Be(46);
-            shape.Y.Should().Be(152);
-            var errors = PptxValidator.Validate(shape.SlideStructure.Presentation);
-            errors.Should().BeEmpty();
-        }
-        
         [Xunit.Theory]
         [SlideShapeData("autoshape-case012.pptx", 1, "Shape 1")]
         public void Text_Setter_should_not_throw_exception(IShape shape)
@@ -296,24 +185,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
             // Assert
             shape.Height.Should().Be(88);
         }
-        
-        [Fact]
-        public void Text_Setter_sets_text_for_New_Shape()
-        {
-            // Arrange
-            var pres = SCPresentation.Create();
-            var shapes = pres.Slides[0].Shapes;
-            var autoShape = shapes.AddRectangle( 50, 60, 100, 70);
-            var textFrame = autoShape.TextFrame!;
-            
-            // Act
-            textFrame.Text = "Test";
-    
-            // Assert
-            textFrame.Text.Should().Be("Test");
-            var errors = PptxValidator.Validate(pres);
-            errors.Should().BeEmpty();
-        }
 
         [Xunit.Theory]
         [SlideShapeData("autoshape-case003.pptx", 1, "AutoShape 6", false)]
@@ -331,25 +202,7 @@ namespace ShapeCrawler.Tests.Unit.xUnit
             // Assert
             textWrapped.Should().Be(isTextWrapped);
         }
-        
-        [Fact]
-        public void AutofitType_Setter_resizes_width()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("autoshape-case003.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var shape = pres.Slides[0].Shapes.GetByName<IAutoShape>("AutoShape 6");
-            var textFrame = shape.TextFrame!;
 
-            // Act
-            textFrame.AutofitType = SCAutofitType.Resize;
-
-            // Assert
-            shape.Width.Should().Be(107);
-            var errors = PptxValidator.Validate(pres);
-            errors.Should().BeEmpty();
-        }
-        
         [Xunit.Theory]
         [SlideShapeData("autoshape-case003.pptx", 1, "AutoShape 7")]
         [SlideShapeData("001.pptx", 1, "Head 1")]
@@ -369,61 +222,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
             errors.Should().BeEmpty();
         }
         
-        [Fact]
-        public void AutofitType_Setter_updates_height()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("autoshape-case003.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var shape = pres.Slides[0].Shapes.GetByName<IAutoShape>("AutoShape 7");
-            var textFrame = shape.TextFrame!;
-
-            // Act
-            textFrame.AutofitType = SCAutofitType.Resize;
-
-            // Assert
-            shape.Height.Should().Be(35);
-            var errors = PptxValidator.Validate(pres);
-            errors.Should().BeEmpty();
-        }
-        
-        [Fact]
-        public void AutofitType_Getter_returns_text_autofit_type()
-        {
-            // Arrange
-            var pptx = GetInputStream("001.pptx");
-            var pres = SCPresentation.Open(pptx);
-            var autoShape = pres.Slides[0].Shapes.GetById<IAutoShape>(9);
-            var textBox = autoShape.TextFrame;
-
-            // Act
-            var autofitType = textBox.AutofitType;
-
-            // Assert
-            autofitType.Should().Be(SCAutofitType.Shrink);
-        }
-
-        [Fact]
-        public void Shape_IsAutoShape()
-        {
-            // Arrange
-            var pres8 = SCPresentation.Open(GetInputStream("008.pptx"));
-            var pres21 = SCPresentation.Open(GetInputStream("021.pptx"));
-            IShape shapeCase1 = SCPresentation.Open(GetInputStream("008.pptx")).Slides[0].Shapes.First(sp => sp.Id == 3);
-            IShape shapeCase2 = SCPresentation.Open(GetInputStream("021.pptx")).Slides[3].Shapes.First(sp => sp.Id == 2);
-            IShape shapeCase3 = SCPresentation.Open(GetInputStream("011_dt.pptx")).Slides[0].Shapes.First(sp => sp.Id == 54275);
-
-            // Act
-            var autoShapeCase1 = shapeCase1 as IAutoShape;
-            var autoShapeCase2 = shapeCase2 as IAutoShape;
-            var autoShapeCase3 = shapeCase3 as IAutoShape;
-
-            // Assert
-            autoShapeCase1.Should().NotBeNull();
-            autoShapeCase2.Should().NotBeNull();
-            autoShapeCase3.Should().NotBeNull();
-        }
-
         [Xunit.Theory]
         [MemberData(nameof(TestCasesParagraphsCount))]
         public void Paragraphs_Count_returns_number_of_paragraphs_in_the_text_box(TestCase<ITextFrame, int> testCase)
@@ -472,79 +270,6 @@ namespace ShapeCrawler.Tests.Unit.xUnit
                 var testCase4 = new TestCase<ITextFrame, int>(4, textBox4, 1);
                 yield return new object[] { testCase4 };
             }
-        }
-
-        [Fact]
-        public void Paragraphs_Add_adds_new_text_paragraph_at_the_end_And_returns_added_paragraph()
-        {
-            // Arrange
-            const string TEST_TEXT = "ParagraphsAdd";
-            var mStream = new MemoryStream();
-            var pres = SCPresentation.Open(GetInputStream("001.pptx"));
-            var textFrame = ((IAutoShape)pres.Slides[0].Shapes.First(sp => sp.Id == 4)).TextFrame;
-            int originParagraphsCount = textFrame.Paragraphs.Count;
-
-            // Act
-            var addedPara = textFrame.Paragraphs.Add();
-            addedPara.Text = TEST_TEXT;
-
-            // Assert
-            var lastPara = textFrame.Paragraphs.Last(); 
-            lastPara.Text.Should().BeEquivalentTo(TEST_TEXT);
-            textFrame.Paragraphs.Should().HaveCountGreaterThan(originParagraphsCount);
-
-            pres.SaveAs(mStream);
-            pres = SCPresentation.Open(mStream);
-            textFrame = ((IAutoShape)pres.Slides[0].Shapes.First(sp => sp.Id == 4)).TextFrame;
-            textFrame.Paragraphs.Last().Text.Should().BeEquivalentTo(TEST_TEXT);
-            textFrame.Paragraphs.Should().HaveCountGreaterThan(originParagraphsCount);
-        }
-
-        [Fact]
-        public void Paragraphs_Add_adds_paragraph()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("autoshape-case007.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var paragraphs = pres.Slides[0].Shapes.GetByName<IAutoShape>("AutoShape 1").TextFrame.Paragraphs;
-            
-            // Act
-            paragraphs.Add();
-            
-            // Assert
-            paragraphs.Should().HaveCount(6);
-        }
-
-        [Fact]
-        public void Paragraphs_Add_adds_new_text_paragraph_at_the_end_And_returns_added_paragraph_When_it_has_been_added_after_text_frame_changed()
-        {
-            var pres = SCPresentation.Open(GetInputStream("001.pptx"));
-            var autoShape = (IAutoShape)pres.Slides[0].Shapes.First(sp => sp.Id == 3);
-            var textBox = autoShape.TextFrame;
-            var paragraphs = textBox.Paragraphs;
-            var paragraph = textBox.Paragraphs.First();
-
-            // Act
-            textBox.Text = "A new text";
-            var newParagraph = paragraphs.Add();
-
-            // Assert
-            newParagraph.Should().NotBeNull();
-        }
-
-        [Fact]
-        public void CanTextChange_returns_false()
-        {
-            // Arrange
-            var pptxStream = GetInputStream("autoshape-case006_field.pptx");
-            var pres = SCPresentation.Open(pptxStream);
-            var textFrame = pres.Slides[0].Shapes.GetByName<IAutoShape>("Field 1").TextFrame;
-            
-            // Act
-            var canTextChange = textFrame.CanChangeText();
-            
-            // Assert
-            canTextChange.Should().BeFalse();
         }
 
         [Xunit.Theory]
@@ -625,5 +350,3 @@ namespace ShapeCrawler.Tests.Unit.xUnit
         }
     }
 }
-
-#endif
